+++ admin/tabs/activity/tracking/ActivityTrackingModel.js
@@ -5,13 +5,14 @@
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
 import {usernameCol} from '@xh/hoist/admin/columns';
+import {RangeAggregator} from '@xh/hoist/admin/tabs/activity/aggregators/RangeAggregator';
 import {ActivityDetailModel} from '@xh/hoist/admin/tabs/activity/tracking/detail/ActivityDetailModel';
 import {GroupingChooserModel} from '@xh/hoist/cmp/grouping';
 import {FilterChooserModel} from '@xh/hoist/cmp/filter';
 import {FormModel} from '@xh/hoist/cmp/form';
 import {GridModel, TreeStyle} from '@xh/hoist/cmp/grid';
 import {HoistModel, managed, XH} from '@xh/hoist/core';
-import {Cube, ChildCountAggregator, LeafCountAggregator, RangeAggregator} from '@xh/hoist/data';
+import {Cube} from '@xh/hoist/data';
 import {fmtDate, fmtNumber, numberRenderer} from '@xh/hoist/format';
 import {action, makeObservable} from '@xh/hoist/mobx';
 import {LocalDate} from '@xh/hoist/utils/datetime';
@@ -83,8 +84,8 @@ export class ActivityTrackingModel extends HoistModel {
                 {name: 'impersonating', type: 'string'},
                 {name: 'dateCreated', displayName: 'Timestamp', type: 'date'},
                 {name: 'data', type: 'json'},
-                {name: 'count', type: 'int', aggregator: new ChildCountAggregator()},
-                {name: 'entryCount', type: 'int', aggregator: new LeafCountAggregator()}
+                {name: 'count', type: 'int', aggregator: 'CHILD_COUNT'},
+                {name: 'entryCount', type: 'int', aggregator: 'LEAF_COUNT'}
             ]
         });
 
@@ -186,7 +187,6 @@ export class ActivityTrackingModel extends HoistModel {
                 {field: 'entryCount', headerName: 'Entries', width: 70, align: 'right'}
             ]
         });
-
 
         this.activityDetailModel = new ActivityDetailModel({parentModel: this});
         this.chartsModel = new ChartsModel({parentModel: this});